pythonProject1/requirements.txt
Appium-Python-Client==4.2.0
pytest==8.3.5
